feat: added quick edits for all media workflows
