+++ services/media/workflows/src/Stations/Collections/CollectionsExplorer/Collections.tsx
@@ -27,6 +27,9 @@ import {
   useUnpublishCollectionMutation,
 } from '../../../generated/graphql';
 import { PublishStatusStateMap } from '../../../Util/PublishStatusStateMap/PublishStatusStateMap';
+import { CollectionDetailsQuickEdit } from '../CollectionDetails/CollectionDetailsQuickEdit';
+import { CollectionEntityManagementQuickEdit } from '../CollectionEntityManagement/CollectionEntityManagementQuickEdit';
+import { CollectionImageManagementQuickEdit } from '../CollectionImageManagement/CollectionImageManagementQuickEdit';
 import { useCollectionsActions } from './Collections.actions';
 import { useCollectionsFilters } from './Collections.filters';
 import { CollectionData } from './Collections.types';
@@ -198,6 +201,22 @@ export const Collections: React.FC = () => {
       filterOptions={filterOptions}
       defaultSortOrder={{ column: 'updatedDate', direction: 'desc' }}
       inlineMenuActions={generateInlineMenuActions}
+      quickEditRegistrations={[
+        {
+          component: <CollectionDetailsQuickEdit />,
+          label: 'Collection Details',
+        },
+        {
+          component: <CollectionEntityManagementQuickEdit />,
+          label: 'Manage Entities',
+          generateDetailsLink: (item) => `/collections/${item.id}/entities`,
+        },
+        {
+          component: <CollectionImageManagementQuickEdit />,
+          label: 'Manage Images',
+          generateDetailsLink: (item) => `/collections/${item.id}/images`,
+        },
+      ]}
     />
   );
 };
