+++ services/media/workflows/src/Stations/Episodes/EpisodesExplorer/Episodes.tsx
@@ -1,5 +1,8 @@
 import React from 'react';
+import { EpisodeDetailsQuickEdit } from '../EpisodeDetails/EpisodeDetailsQuickEdit';
 import { EpisodeExplorer } from '../EpisodeExplorerBase/EpisodeExplorer';
+import { EpisodeImageManagementQuickEdit } from '../EpisodeImageManagement/EpisodeImageManagementQuickEdit';
+import { EpisodeVideoManagementQuickEdit } from '../EpisodeVideoManagement/EpisodeVideoManagementQuickEdit';
 import { useEpisodesActions } from './Episodes.actions';
 
 export const Episodes: React.FC = () => {
@@ -13,6 +16,22 @@ export const Episodes: React.FC = () => {
       bulkActions={bulkActions}
       calculateNavigateUrl={(item) => `/episodes/${item.id}`}
       onCreateAction="/episodes/create"
+      quickEditRegistrations={[
+        {
+          component: <EpisodeDetailsQuickEdit />,
+          label: 'Episode Details',
+        },
+        {
+          component: <EpisodeVideoManagementQuickEdit />,
+          label: 'Manage Videos',
+          generateDetailsLink: (item) => `/episodes/${item.id}/videos`,
+        },
+        {
+          component: <EpisodeImageManagementQuickEdit />,
+          label: 'Manage Images',
+          generateDetailsLink: (item) => `/episodes/${item.id}/images`,
+        },
+      ]}
     />
   );
 };
