+++ services/media/workflows/src/Stations/Movies/MovieExplorerBase/MovieExplorer.tsx
@@ -29,7 +29,6 @@ import {
   useUnpublishMovieMutation,
 } from '../../../generated/graphql';
 import { PublishStatusStateMap } from '../../../Util/PublishStatusStateMap/PublishStatusStateMap';
-import { MovieDetailsQuickEdit } from '../MovieDetails/MovieDetailsQuickEdit';
 import { useMoviesFilters } from './MovieExplorer.filters';
 import { MovieData, MovieExplorerProps } from './MovieExplorer.types';
 
@@ -204,9 +203,6 @@ export const MovieExplorer: React.FC<MovieExplorerProps> = (props) => {
           filterOptions={filterOptions}
           defaultSortOrder={{ column: 'updatedDate', direction: 'desc' }}
           inlineMenuActions={generateInlineMenuActions}
-          quickEditRegistrations={[
-            { component: <MovieDetailsQuickEdit />, label: 'Movie Details' },
-          ]}
         />
       );
     case 'SelectionExplorer':
