+++ services/media/workflows/src/Stations/Movies/MovieImageManagement/MovieImageManagementQuickEdit.tsx
@@ -1,0 +1,13 @@
+import { QuickEditContext, QuickEditContextType } from '@axinom/mosaic-ui';
+import React, { useContext } from 'react';
+import { MovieData } from '../MovieExplorerBase/MovieExplorer.types';
+import { MovieImageManagementForm } from './MovieImageManagementForm';
+
+export const MovieImageManagementQuickEdit: React.FC = () => {
+  const { selectedItem } =
+    useContext<QuickEditContextType<MovieData>>(QuickEditContext);
+
+  return selectedItem ? (
+    <MovieImageManagementForm movieId={selectedItem?.id} />
+  ) : null;
+};
