+++ services/media/workflows/src/Stations/Movies/MovieVideoManagement/MovieVideoManagementQuickEdit.tsx
@@ -1,0 +1,13 @@
+import { QuickEditContext, QuickEditContextType } from '@axinom/mosaic-ui';
+import React, { useContext } from 'react';
+import { MovieData } from '../MovieExplorerBase/MovieExplorer.types';
+import { MovieVideoManagementForm } from './MovieVideoManagementForm';
+
+export const MovieVideoManagementQuickEdit: React.FC = () => {
+  const { selectedItem } =
+    useContext<QuickEditContextType<MovieData>>(QuickEditContext);
+
+  return selectedItem ? (
+    <MovieVideoManagementForm movieId={selectedItem?.id} />
+  ) : null;
+};
