+++ services/media/workflows/src/Stations/Movies/MoviesExplorer/Movies.tsx
@@ -1,5 +1,8 @@
 import React from 'react';
+import { MovieDetailsQuickEdit } from '../MovieDetails/MovieDetailsQuickEdit';
 import { MovieExplorer } from '../MovieExplorerBase/MovieExplorer';
+import { MovieImageManagementQuickEdit } from '../MovieImageManagement/MovieImageManagementQuickEdit';
+import { MovieVideoManagementQuickEdit } from '../MovieVideoManagement/MovieVideoManagementQuickEdit';
 import { useMoviesActions } from './Movies.actions';
 
 export const Movies: React.FC = () => {
@@ -13,6 +16,19 @@ export const Movies: React.FC = () => {
       calculateNavigateUrl={(item) => `/movies/${item.id}`}
       onCreateAction="/movies/create"
       bulkActions={bulkActions}
+      quickEditRegistrations={[
+        { component: <MovieDetailsQuickEdit />, label: 'Movie Details' },
+        {
+          component: <MovieVideoManagementQuickEdit />,
+          label: 'Manage Videos',
+          generateDetailsLink: (item) => `/movies/${item.id}/videos`,
+        },
+        {
+          component: <MovieImageManagementQuickEdit />,
+          label: 'Manage Images',
+          generateDetailsLink: (item) => `/movies/${item.id}/images`,
+        },
+      ]}
     />
   );
 };
