+++ services/media/workflows/src/Stations/Seasons/SeasonImageManagement/SeasonImageManagementQuickEdit.tsx
@@ -1,0 +1,13 @@
+import { QuickEditContext, QuickEditContextType } from '@axinom/mosaic-ui';
+import React, { useContext } from 'react';
+import { SeasonData } from '../SeasonExplorerBase/SeasonExplorer.types';
+import { SeasonImageManagementForm } from './SeasonImageManagementForm';
+
+export const SeasonImageManagementQuickEdit: React.FC = () => {
+  const { selectedItem } =
+    useContext<QuickEditContextType<SeasonData>>(QuickEditContext);
+
+  return selectedItem ? (
+    <SeasonImageManagementForm seasonId={selectedItem?.id} />
+  ) : null;
+};
