+++ services/media/workflows/src/Stations/Seasons/SeasonsExplorer/Seasons.tsx
@@ -1,5 +1,9 @@
 import React from 'react';
+import { SeasonDetailsQuickEdit } from '../SeasonDetails/SeasonDetailsQuickEdit';
+import { SeasonEpisodeManagementQuickEdit } from '../SeasonEpisodeManagement/SeasonEpisodeManagementQuickEdit';
 import { SeasonExplorer } from '../SeasonExplorerBase/SeasonExplorer';
+import { SeasonImageManagementQuickEdit } from '../SeasonImageManagement/SeasonImageManagementQuickEdit';
+import { SeasonVideoManagementQuickEdit } from '../SeasonVideoManagement/SeasonVideoManagementQuickEdit';
 import { useSeasonsActions } from './Seasons.actions';
 
 export const Seasons: React.FC = () => {
@@ -13,6 +17,24 @@ export const Seasons: React.FC = () => {
       bulkActions={bulkActions}
       calculateNavigateUrl={(item) => `/seasons/${item.id}`}
       onCreateAction="/seasons/create"
+      quickEditRegistrations={[
+        { component: <SeasonDetailsQuickEdit />, label: 'Season Details' },
+        {
+          component: <SeasonEpisodeManagementQuickEdit />,
+          label: 'Manage Episodes',
+          generateDetailsLink: (item) => `/seasons/${item.id}/episodes`,
+        },
+        {
+          component: <SeasonVideoManagementQuickEdit />,
+          label: 'Manage Videos',
+          generateDetailsLink: (item) => `/seasons/${item.id}/videos`,
+        },
+        {
+          component: <SeasonImageManagementQuickEdit />,
+          label: 'Manage Images',
+          generateDetailsLink: (item) => `/seasons/${item.id}/images`,
+        },
+      ]}
     />
   );
 };
