+++ services/media/workflows/src/Stations/TvShows/TvShowsExplorer/TvShows.tsx
@@ -1,5 +1,9 @@
 import React from 'react';
+import { TvShowDetailsQuickEdit } from '../TvShowDetails/TvShowDetailsQuickEdit';
 import { TvShowExplorer } from '../TvShowExplorerBase/TvShowExplorer';
+import { TvShowImageManagementQuickEdit } from '../TvShowImageManagement/TvShowImageManagementQuickEdit';
+import { TvShowSeasonManagementQuickEdit } from '../TvShowSeasonManagement/TvShowSeasonManagementQuickEdit';
+import { TvShowVideoManagementQuickEdit } from '../TvShowVideoManagement/TvShowVideoManagementQuickEdit';
 import { useTvShowsActions } from './TvShows.actions';
 
 export const TvShows: React.FC = () => {
@@ -13,6 +17,24 @@ export const TvShows: React.FC = () => {
       calculateNavigateUrl={(item) => `/tvshows/${item.id}`}
       onCreateAction="/tvshows/create"
       bulkActions={bulkActions}
+      quickEditRegistrations={[
+        { component: <TvShowDetailsQuickEdit />, label: 'TV Show Details' },
+        {
+          component: <TvShowSeasonManagementQuickEdit />,
+          label: 'Manage Seasons',
+          generateDetailsLink: (item) => `/tvshows/${item.id}/seasons`,
+        },
+        {
+          component: <TvShowVideoManagementQuickEdit />,
+          label: 'Manage Trailers',
+          generateDetailsLink: (item) => `/tvshows/${item.id}/videos`,
+        },
+        {
+          component: <TvShowImageManagementQuickEdit />,
+          label: 'Manage Images',
+          generateDetailsLink: (item) => `/tvshows/${item.id}/images`,
+        },
+      ]}
     />
   );
 };
